+++ TanksServer/Models/PositionHelpers.cs
@@ -10,15 +10,15 @@ internal static class PositionHelpers
         Debug.Assert(subX < 8);
         Debug.Assert(subY < 8);
         return new PixelPosition(
-            X: position.X * MapService.TileSize + subX,
-            Y: position.Y * MapService.TileSize + subY
+            X: (ushort)(position.X * MapService.TileSize + subX),
+            Y: (ushort)(position.Y * MapService.TileSize + subY)
         );
     }
 
 
     public static PixelPosition ToPixelPosition(this FloatPosition position) => new(
-        X: (int)position.X % MapService.PixelsPerRow,
-        Y: (int)position.Y % MapService.PixelsPerRow
+        X: (ushort)((int)position.X % MapService.PixelsPerRow),
+        Y: (ushort)((int)position.Y % MapService.PixelsPerRow)
     );
 
     public static TilePosition ToTilePosition(this PixelPosition position) => new(
